libbpf-rs: Move link definitions into separate file
